refactor: remove Image, remove Label from Button
Image only did one thing - it dynamically resized to its texture. That functionality was moved to Panel for simplicity. By default, it's disabled, but Label enables it.
The Label on Button is only desired some of the time. The logic was moved to a utility method instead of being a core part of Button.
Updated Panel.resize so the resize event is only fired if the panel size actually changes.