fix: update object sorting
Applying the model-view matrix to the position cancels out the distance. The result can be inconsistent ordering due to float math.
The model's matrices are a composition including the position. The world and view transformations bring it into camera coordinates already, so we can just use the Z translation from the view matrix.

This is an issue mostly for transparent rendering where a transparent mesh in front of another transparent mesh might be rendered in the wrong order and cancel out part of the rendering of the mesh further away.
To be honest, I'm not fully across this. I added a simple demo where the issue is obvious with the old code and not a problem with the change.
There are probably edge cases where this doesn't quite work, but that's always a problem with sorting for rendering.