tweak: add VPosition variable to shaders
The interpolated vertex position for a fragment should be useful.
Made the Position variable in standard shaders a vec3, to match physical shaders.
Removed standard vertex shader in favor of the physical vertex shader which is the same except for texture flips (not sure if that's really needed except for UI).